Web Login Page
This is a web login page that was created using the google cloud app engine. It authenticates the user's login via their google account login information. This projected is coded in python and html.  It is connected to a google cloud server and uses firebase to store data.  This project was created with the help of the google app engine's python 3 tutorials.